added more exceptions

+++ src/main/java/exeptions/ItemAmountUnavailableException.java
@@ -2,7 +2,7 @@ package exeptions;
 
 public class ItemAmountUnavailableException extends Exception {
 
-    int availableAmount;
+    double availableAmount;
     public ItemAmountUnavailableException(String message, Double availableAmount) {
         super(message);
         this.availableAmount = availableAmount;
